refactor(7086): separate binding proof responsibilities
Split task source validation, funding-request verification, settlement orchestration, and report claim assembly into bounded modules so the causal proof path stays reviewable and each new surface remains within repository size limits.

Constraint: Preserve the green task-to-escrow proof contract and canonical verifier behavior.
Rejected: Keep the implementation in runner and one binding module | exceeded file and function limits and mixed validation with writing.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep source validation independent from binding serialization and live funding verification independent from report rendering.
Tested: cargo fmt --check; cargo test -p kamn-e2e-harness; cargo clippy -p kamn-e2e-harness --all-targets --all-features -- -D warnings
Not-tested: workspace-wide gates and funded devnet execution pending integration phase
