+++ crates/kamn-e2e-harness/src/mvp_demo/mod.rs
@@ -18,15 +18,20 @@ mod devnet_settlement_solana;
 mod devnet_settlement_state;
 mod live_task_binding;
 mod live_task_binding_verify;
+mod live_task_funding_verify;
+mod live_task_source_format;
+mod live_task_sources;
 mod local_artifact_paths;
 mod local_artifact_verify;
 mod local_artifacts;
 mod localhost_signed;
 mod report;
 mod report_artifacts;
+mod report_devnet;
 mod report_markdown;
 mod report_writer;
 mod runner;
+mod runner_settlement;
 mod service_api_proof;
 mod three_agent_claim;
 mod three_agent_receipt_spec;
