+++ crates/kamn-e2e-harness/src/mvp_demo/report.rs
@@ -2,18 +2,10 @@ use std::path::Path;
 
 use super::agent_harness::agent_harness_claim_json;
 use super::artifact_digest::ThreeAgentArtifactDigests;
-use super::devnet_settlement::{
-    devnet_no_go_reason, devnet_settlement_claim_json, DevnetSettlementEvidence,
-};
+use super::devnet_settlement::DevnetSettlementEvidence;
 use super::live_task_binding::LiveTaskBinding;
 use super::report_artifacts::artifacts_json;
-use super::three_agent_claim::three_agent_escrow_claim_json;
-use super::three_agent_receipts::{
-    agent_a_observation_receipt_path, agent_b_observation_receipt_path,
-    agent_c_verifier_observation_receipt_path,
-};
-use super::three_agent_transcript::three_agent_transcript_path;
-use super::three_agent_views::{agent_a_view_path, agent_b_view_path, agent_c_verifier_view_path};
+use super::report_devnet::{devnet_required_claims, no_go_json};
 
 #[doc = "MVP demo proof report schema marker."]
 pub const MVP_DEMO_REPORT_SCHEMA_VERSION: &str = "kamn.mvp.demo.proof-report.v1";
@@ -142,75 +134,6 @@ fn roadmap_claim() -> String {
         "{{\"id\":\"production_readiness\",\"label\":\"{}\",\"required\":false,\"status\":\"NOT_CLAIMED\",\"summary\":\"production readiness is not claimed\"}}",
         CLAIM_LABEL_ROADMAP
     )
-}
-
-fn devnet_required_claims(input: &DemoReportInput<'_>) -> Result<Vec<String>, String> {
-    match input.devnet_settlement {
-        Some(evidence) => devnet_success_claims(input, evidence),
-        None => Ok(vec![devnet_no_go_claim_with_reason(input)]),
-    }
-}
-
-fn devnet_success_claims(
-    input: &DemoReportInput<'_>,
-    evidence: &DevnetSettlementEvidence,
-) -> Result<Vec<String>, String> {
-    let settlement_claim = devnet_settlement_claim_json(evidence);
-    let Some(binding) = input.live_task_binding else {
-        return Ok(vec![settlement_claim]);
-    };
-    let digests = input
-        .three_agent_artifact_digests
-        .ok_or_else(|| "missing three-agent artifact digests".to_owned())?;
-    let transcript = three_agent_transcript_path(input);
-    let agent_a = agent_a_view_path(input);
-    let agent_b = agent_b_view_path(input);
-    let agent_c = agent_c_verifier_view_path(input);
-    let agent_a_receipt = agent_a_observation_receipt_path(input);
-    let agent_b_receipt = agent_b_observation_receipt_path(input);
-    let agent_c_receipt = agent_c_verifier_observation_receipt_path(input);
-    Ok(vec![
-        settlement_claim,
-        three_agent_escrow_claim_json(
-            input.run_id,
-            evidence,
-            binding,
-            transcript.as_str(),
-            [agent_a.as_str(), agent_b.as_str(), agent_c.as_str()],
-            [
-                agent_a_receipt.as_str(),
-                agent_b_receipt.as_str(),
-                agent_c_receipt.as_str(),
-            ],
-            digests,
-        ),
-    ])
-}
-
-fn no_go_json(input: &DemoReportInput<'_>) -> String {
-    if input.devnet_mode != "required" || input.devnet_settlement.is_some() {
-        return "{\"active\":false,\"reason\":\"\"}".to_owned();
-    }
-    format!(
-        "{{\"active\":true,\"reason\":\"{}\"}}",
-        effective_no_go_reason(input).as_str()
-    )
-}
-
-fn devnet_no_go_claim_with_reason(input: &DemoReportInput<'_>) -> String {
-    format!(
-        "{{\"id\":\"devnet_settlement_no_go\",\"label\":\"{}\",\"required\":true,\"status\":\"NO-GO\",\"summary\":\"Solana devnet escrow settlement evidence unavailable\",\"network\":\"solana:devnet\",\"rpc_url\":\"{}\",\"no_go_reason\":\"{}\"}}",
-        CLAIM_LABEL_DEVNET_BACKED,
-        escape_json(input.solana_rpc_url.unwrap_or("")),
-        effective_no_go_reason(input).as_str()
-    )
-}
-
-fn effective_no_go_reason(input: &DemoReportInput<'_>) -> String {
-    input
-        .devnet_no_go_reason
-        .unwrap_or_else(|| devnet_no_go_reason(input.solana_rpc_url))
-        .to_owned()
 }
 
 pub(crate) fn escape_json(value: &str) -> String {
